# microservice/source/swagger/swagger.yaml
openapi: 3.0.3
info:
  title: Inventory API
  description: |-
    A demo inventory management system.
  version: 2.0.0
servers:
  - url: https://inventory-api.igraphql.co/api/v2
    description: Microservice
...
paths:
...
  /inventory/store/{storeId}:
    delete:
      tags:
        - Inventory
      description: Deletes a store's inventory.
      parameters:
        - name: storeId
          in: path
          description: Store ID.
          required: true
          schema:
            type: integer
            example: 6602
      responses:
        "200":
          description: Inventory successfully deleted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode200"
        "404":
          description: Store not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode404"
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode500"
...
  /stores:
    get:
      tags:
        - Stores
      description: Retrieves stores.
      parameters:
        - $ref: '#/components/parameters/Address'
        - $ref: '#/components/parameters/City'
        - $ref: '#/components/parameters/Field'
        - $ref: '#/components/parameters/Latitude'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Longitude'
        - $ref: '#/components/parameters/NameStore'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/State'
        - $ref: '#/components/parameters/StoreID'
        - $ref: '#/components/parameters/ZipCode'
      responses:
        "200":
          description: Store(s) successfully retrieved.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetStores"
        "404":
          description: Store(s) not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode404"
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode500"
  /stores/{storeId}:
    post:
      tags:
        - Stores
      description: Creates a store.
      parameters:
        - name: storeId
          in: path
          description: Store ID.
          required: true
          schema:
            type: integer
            example: 3025
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStore'
      responses:
        "201":
          description: Store successfully created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode201"
        "400":
          description: Client sent an invalid request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode400"
        "409":
          description: Store already exists.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode409"
        "415":
          description: Invalid Content-Type.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode415"
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode500"
    get:
      tags:
        - Stores
      description: Retrieves a store.
      parameters:
        - name: storeId
          in: path
          description: Store ID.
          required: true
          schema:
            type: integer
            example: 3025
        - $ref: '#/components/parameters/Field'
      responses:
        "200":
          description: Store successfully retrieved.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetStore"
        "404":
          description: Store not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode404"
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode500"
    put:
      tags:
        - Stores
      description: Updates a store.
      parameters:
        - name: storeId
          in: path
          description: Store ID.
          required: true
          schema:
            type: integer
            example: 3025
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStore'
      responses:
        "200":
          description: Store successfully updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode200"
        "400":
          description: Client sent an invalid request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode400"
        "404":
          description: Store not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode404"
        "415":
          description: Invalid Content-Type.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode415"
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode500"
    delete:
      tags:
        - Stores
      description: Deletes a store.
      parameters:
        - name: storeId
          in: path
          description: Store ID.
          required: true
          schema:
            type: integer
            example: 3025
      responses:
        "200":
          description: Store successfully deleted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode200"
        "404":
          description: Store not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode404"
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode500"
components:
  parameters:
    Address:
      name: address
      in: query
      description: Store address.
      schema:
        type: string
...
    City:
      name: city
      in: query
      description: Store city.
      schema:
        type: string
...
    Field:
      name: field
      in: query
      description: Limit the fields returned by the query. Can be specified multiple times.
      schema:
        type: array
        items:
          type: string
      style: form
      explode: true
    Latitude:
      name: latitude
      in: query
      description: Store latitude.
      schema:
        type: number
    Limit:
      name: limit
      in: query
      description: Limit the number of values returned by the query.
      schema:
        type: integer
    Longitude:
      name: longitude
      in: query
      description: Store longitude.
      schema:
        type: number
...
    NameStore:
      name: name
      in: query
      description: Store name.
      schema:
        type: string
    Offset:
      name: offset
      in: query
      description: Offset the number of values returned by the query.
      schema:
        type: integer
...
    State:
      name: state
      in: query
      description: Store state.
      schema:
        type: string
    StoreID:
      name: store_id
      in: query
      description: Store ID.
      schema:
        type: integer
...
    ZipCode:
      name: zip_code
      in: query
      description: Store zip code.
      schema:
        type: integer
  schemas:
...
    CreateStore:
      type: object
      required:
        - address
        - city
        - latitude
        - longitude
        - name
        - state
        - zip_code
      properties:
        address:
          type: string
          example: 1247 West El Camino Real
        city:
          type: string
          example: Sunnyvale
        latitude:
          type: number
          example: 37.375004
        longitude:
          type: number
          example: -122.057389
        name:
          type: string
          example: Sunnyvale
        state:
          type: string
          example: CA
        zip_code:
          type: integer
          example: 94086
    GetStore:
      type: object
      properties:
        address:
          type: string
          example: 1247 West El Camino Real
        city:
          type: string
          example: Sunnyvale
        created:
          type: string
          example: 2023-07-04T12:00:00.00.000Z
        latitude:
          type: number
          example: 37.375004
        longitude:
          type: number
          example: -122.057389
        name:
          type: string
          example: Sunnyvale
        state:
          type: string
          example: CA
        store_id:
          type: integer
          example: 3025
        updated:
          type: string
          example: 2023-07-04T12:00:00.00.000Z
        zip_code:
          type: integer
          example: 94086
    GetStores:
      type: array
      items:
        type: object
        properties:
          address:
            type: string
            example: 1247 West El Camino Real
          city:
            type: string
            example: Sunnyvale
          created:
            type: string
            example: 2023-07-04T12:00:00.00.000Z
          latitude:
            type: number
            example: 37.375004
          longitude:
            type: number
            example: -122.057389
          name:
            type: string
            example: Sunnyvale
          state:
            type: string
            example: CA
          store_id:
            type: integer
            example: 3025
          updated:
            type: string
            example: 2023-07-04T12:00:00.00.000Z
          zip_code:
            type: integer
            example: 94086
    UpdateStore:
      type: object
      properties:
        address:
          type: string
          example: 1247 West El Camino Real
        city:
          type: string
          example: Sunnyvale
        latitude:
          type: number
          example: 37.375004
        longitude:
          type: number
          example: -122.057389
        name:
          type: string
          example: Sunnyvale
        state:
          type: string
          example: CA
        zip_code:
          type: integer
          example: 94086
    HttpCode200:
      type: object
      properties:
        status:
          type: integer
          example: 200
        message:
          type: string
          example: ok
    HttpCode201:
      type: object
      properties:
        status:
          type: integer
          example: 201
        message:
          type: string
          example: created
    HttpCode400:
      type: object
      properties:
        status:
          type: integer
          example: 400
        message:
          type: string
          example: bad request
    HttpCode404:
      type: object
      properties:
        status:
          type: integer
          example: 404
        message:
          type: string
          example: not found
    HttpCode409:
      type: object
      properties:
        status:
          type: integer
          example: 409
        message:
          type: string
          example: conflict
    HttpCode415:
      type: object
      properties:
        status:
          type: integer
          example: 415
        message:
          type: string
          example: unsupported media type
    HttpCode500:
      type: object
      properties:
        status:
          type: integer
          example: 500
        message:
          type: string
          example: internal server error